Initial commit of reviewer-by-blame plugin
A plugin that allows to automatically add reviewers to a change
from the git blame computation on the changed files. It will add
the users as reviewer that authored most of the lines touched by
the change, since these users should be familiar with the code
and can mostly review the change.

Change-Id: I20df7ffcefdd88189238c637ae301e0d3e4ef796