Support configuration of max reviewers per project
It is now possible to define the maximum number of reviewers that
should be added to a change by this plugin on project level in the
'project.config' file of the project. If not defined on a project
the value is inherited from the parent projects. The default value is
3.

Change-Id: If5b1a56ae3df964154bc6c6b2487ada7fea49650
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>