Do not fail if patch set is not found
If a change ref is updated the plugin expects to find the
corresponding patch set in the database and fails if the patch set
cannot be found in the database.

Add a check for this case and do not fail, but instead ignore this ref
update and write a warning into the log.

This error happens if a change is rebased from the Gerrit WebUI
because Gerrit sends the ref-updated event for the change before it
creates the new patch set in the database.

Change-Id: Ie79fd5abace82a98bbc499031403b90443dcdde7
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>