Update patch set 1
Patch Set 1:

Yeah, how about to rename that plugin (or delete and create with a new name, because i guess we can not still rename projects?) to be simple

 reviewers-plugin

or some such? I can think of another strategy: like "peek random reviewers" strategy from a list of reviewers.

I feel very wrong to have dedicated plugins for every (new) strategy. So let's abuse this one ;-)

And the best is: we can do it per project base. So each project decides which automatic reviewers assigning strategy it wants, if at all.

Patch-set: 1