Update patch set 1
Patch Set 1:

Is there a problem with keeping the pom.xml?

I kind of like building plugins standalone with Maven. With Buck it may use a stale version of the plugin API and some Buck modifications as described in [1] are needed to avoid this issue... with Maven I always build against my latest API version.

[1] https://gerrit-review.googlesource.com/Documentation/dev-buck.html#_building_against_unpublished_jars_that_change_frequently

Patch-set: 1