Update patch set 1
Patch Set 1: Code-Review+1

> I was removing it for consistency with the others, that don't have
 > it (except those that currently can't be built with buck).

There are quite some plugins which can be built with Buck and Maven. I'm okay with removing the pom.xml, but on the other hand it's also not disturbing much.

 > How does building with Maven help with the API version?  You still
 > have to build the API locally.

If I build and install the 2.11-SNAPSHOT API version, then with Maven always this latest API version is used. With Buck it is retrieved once and then cached. Any new version of 2.11-SNAPSHOT API which I build and install afterwards is not seen, and I must do this adaption to make it work. 

 > Using an official API is also possible with BUCK, like in [1]
 > 
 > [1] https://gerrit-review.googlesource.com/64681

That's cool :)
Maybe we need to do this for more plugins.

Patch-set: 1
Label: Code-Review=+1