Update patch set 21
Patch Set 21:

Ideally as a user I'd hope to have a user settings option for not being added as a reviewer, instead of having to ask the project admin of every project I contributed to to add me to a blacklist. (Typically this affects people who are not active anymore, or do commits on a lot of different repos where they don't have the knowledge or bandwidth to do reviews (e.g. they are doing semi-automated code style fixes), or are relatively new to Gerrit and still unsure what's happening and why. For different reasons but in each of those three cases forcing the user to talk to repo admins would be hard on them.)

Patch-set: 21
CC: Gerrit User 1063876 <1063876@173816e5-2b9a-37c3-8a2e-48639d4f1153>