Update patch set 28
Patch Set 28:

> Patch Set 28:
> 
> > Please consider this patchset as a single isolated change. These comments are too broad to be addressed in a single patch.
>
> I don't get how this is a response to what I wrote. I explained why turning the process as implemented by this plugin into a button makes it worse. [...]

I am not following the recent requests for adjusting this plugin for Wikimedia needs, but I would really like to take unrealted discussion away from comment thread of single changes. There are other means, like issue tracker or mailing lists for that.

For example this comment on this change is unrelated to this change and is tracked here: [1].

   Will this still misattribute the request emails to the original uploader of the patch?

So either discuss it there in issue tracker, or fix it in its own change and discuss it there.

* [1] https://bugs.chromium.org/p/gerrit/issues/detail?id=10328

Patch-set: 28
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>