Adapth the code to core on stable-3.5 branch
In I81eeadb9fd the code was adapted to DiffOperations, but this
adaptation was already done against I22a465e49 that was included
starting from stable-3.6 branch.

Remove DiffOptions parameter to the invocation of:
DiffOperations#listModifiedFilesAgainstParent(). Also note, that
this change should be reverted when merging it to stable-3.6 branch.

Change-Id: I0f15022efd6322a44e38cf452bd060e0be60c510