Add better handling of private and WIP changes.
Since private changes cannot be seen by anyone, do not add reviewers to
them. Further, add an option (defaulting to true) to not add reviewers
to WIP changes. As soon as the change is marked ready for review (either
from WIP or from Private), reviewers will be added.

This is nicer than a subject line since gerrit already offers this
functionality natively.

Change-Id: I0a3c02551249e8401cf7a0f545335d2bc6c8d9cd